+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - integrated log system
-- token to subscription 
+- token to subscription
+- custom actions support
 
 ### Changed